+++ src/components/navigation/Navigation/BottomNavigationAction.stories.tsx
@@ -8,7 +8,7 @@ import FavoriteIcon from '@material-ui/icons/Favorite'
 import LocationOnIcon from '@material-ui/icons/LocationOn'
 
 export default {
-  title: '💧 Atoms/BottomNavigationAction',
+  title: 'Navigation/BottomNavigationAction',
   component: BottomNavigationAction,
   args: {
     label: {
