+++ src/components/navigation/index.ts
@@ -2,8 +2,14 @@ import { Link } from './Link/Link'
 import { MenuItem } from './Menu/MenuItem'
 import { Menu } from './Menu/Menu'
 import { Breadcrumbs } from './Breadcrumbs/Breadcrumbs'
-import { BottomNavigation } from './BottomNavigation/BottomNavigation'
-import { BottomNavigationAction } from './BottomNavigationAction/BottomNavigationAction'
+import { BottomNavigation } from './Navigation/BottomNavigation'
+import { BottomNavigationAction } from './Navigation/BottomNavigationAction'
 
-
-export { Link, MenuItem, Menu, Breadcrumbs, BottomNavigation, BottomNavigationAction }
+export {
+  Link,
+  MenuItem,
+  Menu,
+  Breadcrumbs,
+  BottomNavigation,
+  BottomNavigationAction
+}
